feat: consumer api 수정 및 분리
